chore: pull latest changes and resolve conflicts
