+++ messages/context.json
@@ -11,6 +11,9 @@
   "admin/glovo-integration.table.edit-store": "Edit store",
   "admin/glovo-integration.table.remove-store": "Remove store",
   "admin/glovo-integration.table.empty": "There are no stores configured",
+  "admin/glovo-integration.table.show-rows": "Show rows",
+  "admin/glovo-integration.table.of": "of",
+  "admin/glovo-integration.collapsible.header": "Configuration",
   "admin/glovo-integration.integration-settings.title": "Integration Settings",
   "admin/glovo-integration.integration-settings.subtitle": "Glovo integration settings",
   "admin/glovo-integration.integration-settings.glovo-token.description": "Token to perform calls to the Glovo API",
