+++ react/components/admin/Stores.tsx
@@ -1,26 +1,21 @@
 import type { FC } from 'react'
 import React, { useState, Fragment } from 'react'
 import { FormattedMessage } from 'react-intl'
-import {
-  ButtonWithIcon,
-  ModalDialog,
-  Table,
-  IconPlusLines,
-} from 'vtex.styleguide'
+import { ButtonWithIcon, ModalDialog, IconPlusLines } from 'vtex.styleguide'
 
 import { IconGlovo } from '../../icons/IconGlovo'
-import { countries } from '../../utils'
 import { StoreModal } from './StoreModal'
 import type { AlertProps } from './AlertBanner'
 import { AlertBanner } from './AlertBanner'
 import type { SettingsSection } from '../../typings/settingsSections'
+import { StoreTable } from './StoresTable'
 
 interface ModifyStore {
   isOpen: boolean
   loading: boolean
 }
 
-interface RemoveStore extends ModifyStore {
+export interface RemoveStore extends ModifyStore {
   storeId: string
 }
 
@@ -136,62 +131,6 @@ const Stores: FC<SettingsSection> = ({ settings, saveSettings }) => {
     })
   }
 
-  const schema = {
-    properties: {
-      storeName: {
-        title: <FormattedMessage id="admin/glovo-integration.store-name" />,
-      },
-      affiliateId: {
-        title: <FormattedMessage id="admin/glovo-integration.affiliate-id" />,
-      },
-      salesChannel: {
-        title: <FormattedMessage id="admin/glovo-integration.sales-channel" />,
-      },
-      glovoStoreId: {
-        title: <FormattedMessage id="admin/glovo-integration.glovo-store-id" />,
-      },
-      postalCode: {
-        title: <FormattedMessage id="admin/glovo-integration.postal-code" />,
-      },
-      country: {
-        title: <FormattedMessage id="admin/glovo-integration.country" />,
-        cellRenderer: ({ cellData }: TableItem) => {
-          const country = countries.find(({ value }) => value === cellData)
-
-          return country?.label
-        },
-      },
-    },
-  }
-
-  const lineActions = [
-    {
-      label: () => (
-        <FormattedMessage id="admin/glovo-integration.table.edit-store" />
-      ),
-      onClick: ({ rowData: store }: TableItem) => {
-        setAddOrEditStore({
-          ...addOrEditStore,
-          isOpen: true,
-          store,
-        })
-      },
-    },
-    {
-      label: () => (
-        <FormattedMessage id="admin/glovo-integration.table.remove-store" />
-      ),
-      isDangerous: true,
-      onClick: ({ rowData: { id } }: TableItem) => {
-        setRemoveStore({
-          ...removeStore,
-          isOpen: true,
-          storeId: id,
-        })
-      },
-    },
-  ]
-
   return (
     <Fragment>
       <StoreModal
@@ -245,19 +184,12 @@ const Stores: FC<SettingsSection> = ({ settings, saveSettings }) => {
           </ButtonWithIcon>
         </div>
 
-        <Table
-          schema={schema}
+        <StoreTable
           items={settings.stores}
-          lineActions={lineActions}
-          fullWidth
-          emptyStateLabel=""
-          emptyStateChildren={
-            <div className="">
-              <h4 className="t-heading-4 pt8">
-                <FormattedMessage id="admin/glovo-integration.table.empty" />
-              </h4>
-            </div>
-          }
+          addOrEditStore={addOrEditStore}
+          setAddOrEditStore={setAddOrEditStore}
+          removeStore={removeStore}
+          setRemoveStore={setRemoveStore}
         />
       </div>
     </Fragment>
